add powi to java

+++ java/libsvm/svm_parameter.java
@@ -16,7 +16,7 @@ public class svm_parameter implements Cloneable,java.io.Serializable
 
 	public int svm_type;
 	public int kernel_type;
-	public double degree;	// for poly
+	public int degree;	// for poly
 	public double gamma;	// for poly/rbf/sigmoid
 	public double coef0;	// for poly/sigmoid
