+++ java/svm_toy.java
@@ -201,7 +201,7 @@ public class svm_toy extends Applet {
 					param.kernel_type = atoi(argv[i]);
 					break;
 				case 'd':
-					param.degree = atof(argv[i]);
+					param.degree = atoi(argv[i]);
 					break;
 				case 'g':
 					param.gamma = atof(argv[i]);
